Add breakpoint and cleanup path mess

+++ Installer_Src/Setup DreamWorld/FileStuff.vb
@@ -12,9 +12,10 @@ Module FileStuff
 
         Try
             My.Computer.FileSystem.CopyFile(Source, Dest, overwrite)
-#Disable Warning CA1031
+
         Catch ex As Exception
-#Enable Warning CA1031
+
+            BreakPoint.Show(ex.Message)
         End Try
 
     End Sub
@@ -27,9 +28,10 @@ Module FileStuff
         If Not System.IO.Directory.Exists(destinationPath) Then
             Try
                 System.IO.Directory.CreateDirectory(destinationPath)
-#Disable Warning CA1031
+
             Catch ex As Exception
-#Enable Warning CA1031
+
+                BreakPoint.Show(ex.Message)
             End Try
         End If
 
@@ -52,18 +54,20 @@ Module FileStuff
 
                 Try
                     CopyFile(fileSystemInfo.FullName, destinationFileName, True)
-#Disable Warning CA1031
+
                 Catch ex As Exception
-#Enable Warning CA1031
+
+                    BreakPoint.Show(ex.Message)
                 End Try
             Else
                 ' Recursively call the method to copy all the nested folders
                 If Not System.IO.Directory.Exists(fileSystemInfo.FullName) Then
                     Try
                         System.IO.Directory.CreateDirectory(fileSystemInfo.FullName)
-#Disable Warning CA1031
+
                     Catch ex As Exception
-#Enable Warning CA1031
+
+                        BreakPoint.Show(ex.Message)
                     End Try
                 End If
                 CopyFolder(fileSystemInfo.FullName, destinationFileName)
@@ -76,20 +80,18 @@ Module FileStuff
     Sub DeleteDirectory(folder As String, param As FileIO.DeleteDirectoryOption)
         Try
             My.Computer.FileSystem.DeleteDirectory(folder, param)
-#Disable Warning CA1031
+
         Catch ex As Exception
-#Enable Warning CA1031
+
+            BreakPoint.Show(ex.Message)
         End Try
     End Sub
 
     Sub DeleteFile(file As String)
 
-        Try
+        If My.Computer.FileSystem.FileExists(file) Then
             My.Computer.FileSystem.DeleteFile(file)
-#Disable Warning CA1031
-        Catch ex As Exception
-#Enable Warning CA1031
-        End Try
+        End If
 
     End Sub
